Ticket [https://trac.osgeo.org/mapguide/ticket/1295] Improve build scripts for 2.2 release cycle Modify linux_apt.sh to check for apache2 directory before attempting stop.
git-svn-id: http://svn.osgeo.org/mapguide/trunk@4668 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Oem/LinuxApt/build_apt.sh
@@ -134,16 +134,19 @@ echo Apache Httpd build completed
 # Apache shutdown  procedure
 # Notes: none
 #**********************************************************
-echo Attempting to shutdown Apache
-pushd $INSTALLWEB/apache2/bin
-./apachectl stop
-popd
-echo Attempting to remove old Apache and Php
-pushd $INSTALLWEB
-rm -rf apache2
-rm -rf php
-popd
-echo Completed uninstall of Apache and Php
+echo Checking for existing Apache install
+if [ -d $INSTALLWEB/apache2/bin ]; then
+  echo Attempting to shutdown Apache
+  pushd $INSTALLWEB/apache2/bin
+  ./apachectl stop
+  popd
+  echo Attempting to remove old Apache and Php
+  pushd $INSTALLWEB
+  rm -rf apache2
+  rm -rf php
+  popd
+  echo Completed uninstall of Apache and Php
+fi
 
 #**********************************************************
 # Apache install procedure
